+++ examples/microbenchmark/bw_weak/arl_agg_bw.cpp
@@ -57,10 +57,10 @@ void worker() {
   print("req payload size is %d Byte\n", REQ_N);
   print("ack payload size is %d Byte\n", ACK_N);
   print("rpc_agg overhead is %.2lf us (total %.2lf s)\n", duration_total / num_ops, duration_total / 1e6);
-  print("Total single-direction node bandwidth (req/gross): %lu MB/s\n", (unsigned long) ((sizeof(amagg_internal::AmaggReqMeta) + sizeof(ReqPayload<REQ_N>)) * num_ops * local::rank_n() * 2 / duration_total));
-  print("Total single-direction node bandwidth (req/pure): %lu MB/s\n", (unsigned long) ((sizeof(ReqPayload<REQ_N>)) * num_ops * local::rank_n() * 2 / duration_total));
-  print("Total single-direction node bandwidth (ack/gross): %lu MB/s\n", (unsigned long) ((sizeof(amagg_internal::AmaggAckMeta) + sizeof(AckPayload<ACK_N>)) * num_ops * local::rank_n() * 2 / duration_total));
-  print("Total single-direction node bandwidth (ack/pure): %lu MB/s\n", (unsigned long) ((sizeof(AckPayload<ACK_N>)) * num_ops * local::rank_n() * 2 / duration_total));
+  print("Total single-direction node bandwidth (req/gross): %.2lf MB/s\n", ((sizeof(amagg_internal::AmaggReqMeta) + sizeof(ReqPayload<REQ_N>)) * num_ops * local::rank_n() * 2 / duration_total));
+  print("Total single-direction node bandwidth (req/pure): %.2lf MB/s\n", ((sizeof(ReqPayload<REQ_N>)) * num_ops * local::rank_n() * 2 / duration_total));
+  print("Total single-direction node bandwidth (ack/gross): %.2lf MB/s\n", ((sizeof(amagg_internal::AmaggAckMeta) + sizeof(AckPayload<ACK_N>)) * num_ops * local::rank_n() * 2 / duration_total));
+  print("Total single-direction node bandwidth (ack/pure): %.2lf MB/s\n", ((sizeof(AckPayload<ACK_N>)) * num_ops * local::rank_n() * 2 / duration_total));
 }
 
 int main(int argc, char** argv) {
