+++ examples/microbenchmark/bw_weak/arl_ff_bw.cpp
@@ -39,8 +39,8 @@ void worker() {
   double duration_total = ticks_to_us(end_wait - start);
   print("req payload size is %d Byte\n", N);
   print("rpc_ff overhead is %.2lf us (total %.2lf s)\n", duration_total / num_ops, duration_total / 1e6);
-  print("Total single-direction node bandwidth (req/gross): %lu MB/s\n", (unsigned long) ((sizeof(amff_internal::AmffReqMeta) + sizeof(Payload<N>)) * num_ops * local::rank_n() * 2 / duration_total));
-  print("Total single-direction node bandwidth (req/pure): %lu MB/s\n", (unsigned long) ((sizeof(Payload<N>)) * num_ops * local::rank_n() * 2 / duration_total));
+  print("Total single-direction node bandwidth (req/gross): %.2lf MB/s\n", ((sizeof(amff_internal::AmffReqMeta) + sizeof(Payload<N>)) * num_ops * local::rank_n() * 2 / duration_total));
+  print("Total single-direction node bandwidth (req/pure): %.2lf MB/s\n", ((sizeof(Payload<N>)) * num_ops * local::rank_n() * 2 / duration_total));
 }
 
 int main(int argc, char** argv) {
